+++ integration-tests/worker/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @openfn/integration-tests-worker
 
+## 1.0.37
+
+### Patch Changes
+
+- @openfn/engine-multi@1.1.2
+- @openfn/lightning-mock@2.0.2
+- @openfn/ws-worker@1.1.2
+
 ## 1.0.36
 
 ### Patch Changes